+++ src/views/partner/managehotel/ManageRoom.vue
@@ -156,6 +156,7 @@ import Swal from "sweetalert2";
 import EditHotel from "./EditHotel.vue";
 import Gallery from "../../../components/Gallery.vue";
 import Loading from "../../../components/Loading.vue";
+
 export default {
   components: {
     EditHotel,
@@ -168,6 +169,7 @@ export default {
   },
 
   setup() {
+    const Modal = ref(false);
     const searchall = ref("");
     const item_product = ref([]);
     const reservationEnabled = ref(false);
@@ -204,28 +206,48 @@ export default {
 
     const deleteProduct = async (_id) => {
       try {
-        const response = await axios.delete(
-          `${process.env.VUE_APP_API}room/${_id}`,
-          {
-            headers: {
-              token: localStorage.getItem("token"),
-            },
+        const { value: deletehotel } = await Swal.fire({
+          title: "ต้องการลบหรือไม่",
+          text: "คุณต้องการลบรายการนี้หรือไม่",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          cancelButtonText: "ยกเลิก",
+          confirmButtonText: "ยืนยันลบ",
+          // input: "text",
+          // inputLabel: "สาเหตุที่ต้องการลบ",
+          // inputPlaceholder: "กรุณากรอกสาเหตุ",
+          // inputValidator: (value) => {
+          //   if (!value) {
+          //     return "กรุณากรอกสาเหตุที่ต้องการลบ";
+          //   }
+          // },
+        });
+
+        if (deletehotel) {
+          const response = await axios.delete(
+            `${process.env.VUE_APP_API}room/${_id}`,
+            {
+              headers: {
+                token: localStorage.getItem("token"),
+              },
+            }
+          );
+
+          if (response.data) {
+            getData();
+            await Swal.fire({
+              icon: "success",
+              title: "ลบรายการสำเร็จ",
+            });
+          } else {
+            await Swal.fire({
+              icon: "error",
+              title: "เกิดข้อผิดพลาด",
+              text: "ไม่สามารถแก้ไขข้อมูลได้",
+            });
           }
-        );
-        if (response.data) {
-          // หากการลบสำเร็จ อัปเดตข้อมูล
-          getData();
-          // แสดงข้อความสำเร็จ (ตัวเลือก)
-          Swal.fire({
-            icon: "success",
-            title: "ลบรายการสำเร็จ",
-          });
-        } else {
-          await Swal.fire({
-            icon: "error",
-            title: "เกิดข้อผิดพลาด",
-            text: "ไม่สามารถแก้ไขข้อมูลได้",
-          });
         }
       } catch (error) {
         await Swal.fire({
@@ -308,6 +330,7 @@ export default {
       statusdata,
       selectstatus,
       headerTextColor: "rgb(156 163 175)",
+      Modal,
     };
   },
